requirements.txt
# ─── Core Django ──────────────────────────────────────────────────────────────
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-axes==6.5.1
django-filter==24.2
django-redis==5.4.0

# ─── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9
django-extensions==3.2.3

# ─── Task queue ───────────────────────────────────────────────────────────────
celery==5.4.0
redis==5.0.8
django-celery-beat==2.6.0
django-celery-results==2.5.1

# ─── FIDO2/WebAuthn ───────────────────────────────────────────────────────────
fido2==1.1.3

# ─── Cryptography ─────────────────────────────────────────────────────────────
cryptography==43.0.0
PyNaCl==1.5.0

# ─── Payment providers ────────────────────────────────────────────────────────
stripe==10.7.0
requests==2.32.3

# ─── IPFS ─────────────────────────────────────────────────────────────────────
ipfshttpclient==0.8.0a2

# ─── Utilities ────────────────────────────────────────────────────────────────
python-decouple==3.8
Pillow==10.4.0
python-jose[cryptography]==3.3.0
pydantic==2.8.2
structlog==24.4.0
sentry-sdk[django]==2.13.0

# ─── GDPR / compliance ────────────────────────────────────────────────────────
django-gdpr-assist==1.5.0
